# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Api Juegos",
        "description": "Documentación de todos los endpoints de un aplicación que consiste en un CRUD de Juegos que contiene relaciones, donde se trabaja con el envió de archivos y también contiene filtros de búsqueda sobre resultados de los juegos que existen insertados, se puede registrar usuarios y también contiene autenticación por SANCTUM para poder realizar la acción de algunos endpoints.<br><br><a href='https://github.com/JAVI-CC/Laravel-API-Server' target='_blank'>https://github.com/JAVI-CC/Laravel-API-Server</a><br><br><a href='https://github.com/JAVI-CC/Laravel-API-Client' target='_blank'>https://github.com/JAVI-CC/Laravel-API-Client</a>",
        "version": "1.0.1"
    },
    "paths": {
...
        "/api/juegos/desarrolladoras/{slug}": {
            "get": {
                "tags": [
                    "Desarrolladoras"
                ],
                "summary": "Obtener los juegos de una desarrolladora",
                "description": "Muestra la información de los juegos que pertenece de una desarrolladora en especifico segun el valor del parametro slug.",
                "operationId": "getDesarrolladora",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Url del nombre de la desarrolladora",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "test-studios"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "SANCTUM": {
                "type": "apiKey",
                "description": "Clave API de tipo Bearer Token para poder utilizar todas las peticiones.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}